Extract handlers from SectionHotspot
JIRA: RAIL-4170
